pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "finance-alerts"
version = "0.1.0"
description = "Finance news impact scoring and Telegram alert SaaS"
requires-python = ">=3.12"
dependencies = [
    # Web framework
    "fastapi>=0.111.0",
    "uvicorn[standard]>=0.30.0",
    # Task queue
    "celery[redis]>=5.4.0",
    # Database
    "sqlalchemy[asyncio]>=2.0.30",
    "asyncpg>=0.29.0",
    "alembic>=1.13.0",
    # Redis
    "redis>=5.0.0",
    # HTTP client
    "httpx>=0.27.0",
    # RSS parsing
    "feedparser>=6.0.11",
    # Web scraping fallback
    "trafilatura>=1.9.0",
    # Deduplication
    "rapidfuzz>=3.9.0",
    # Language detection
    "langdetect>=1.0.9",
    # Configuration
    "pydantic-settings>=2.3.0",
    "pyyaml>=6.0.1",
    # Stripe
    "stripe>=10.5.0",
    # Logging
    "structlog>=24.2.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.2.0",
    "pytest-asyncio>=0.23.0",
    "pytest-cov>=5.0.0",
    "httpx>=0.27.0",
    "factory-boy>=3.3.0",
    "ruff>=0.4.0",
    "mypy>=1.10.0",
]

[tool.ruff]
line-length = 100
target-version = "py312"

[tool.ruff.lint]
select = ["E", "F", "I", "UP"]

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]

[tool.mypy]
python_version = "3.12"
strict = false
ignore_missing_imports = true
